# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 商机随手报 API
  version: 1.0.0
  description: 固定问卷、区县承接、分侧处理、双节点审核与市公司成功商机库。
servers:
  - url: /api/v1
security:
  - cookieAuth: []
    roleGrant: []
paths:
...
  /auth/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "201":
          description: logged in
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/CurrentUser" } },
            }
        default: { $ref: "#/components/responses/Problem" }
  /auth/me:
    get:
      security: [{ cookieAuth: [] }]
      operationId: currentUser
      responses:
        "200":
          {
            description: current user,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/CurrentUser" } },
              },
          }
        default: { $ref: "#/components/responses/Problem" }
  /auth/change-password:
    post:
      security: [{ cookieAuth: [] }]
      operationId: changePassword
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: "#/components/schemas/ChangePasswordRequest" },
                },
            },
        }
      responses:
        "201": { description: changed }
        default: { $ref: "#/components/responses/Problem" }
  /auth/reauth:
    post:
      operationId: reauthenticateForExport
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/ReauthRequest" } },
            },
        }
      responses:
        "201":
          {
            description: one-time token,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: object,
                        additionalProperties: false,
                        required: [token, expiresAt],
                        properties:
                          {
                            token: { type: string },
                            expiresAt: { type: string, format: date-time },
                          },
                      },
                  },
              },
          }
        default: { $ref: "#/components/responses/Problem" }
  /auth/logout:
    post:
      security: [{ cookieAuth: [] }]
      operationId: logout
      responses:
        "201": { description: logged out }
...
  /reference/handlers:
    get:
      operationId: listAssignableHandlers
      parameters:
        - {
            in: query,
            name: districtId,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            in: query,
            name: customerType,
            required: true,
            schema: { $ref: "#/components/schemas/CustomerType" },
          }
      responses:
        "200": { description: handlers }
        default: { $ref: "#/components/responses/Problem" }
...
  /opportunities/{opportunityId}/audio:
    parameters: [{ $ref: "#/components/parameters/OpportunityId" }]
    post:
      operationId: uploadAudio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [audio]
              properties: { audio: { type: string, format: binary } }
      responses:
        "201": { description: encrypted audio stored }
        default: { $ref: "#/components/responses/Problem" }
    get:
      operationId: playAudio
      responses:
        "200":
          {
            description: decrypted authorized stream,
            content:
              {
                audio/webm: { schema: { type: string, format: binary } },
                audio/mp4: { schema: { type: string, format: binary } },
              },
          }
        default: { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteAudioBeforeReview
      responses:
        "200": { description: reporter removed audio while editable }
        default: { $ref: "#/components/responses/Problem" }
...
  /municipal/successes:
    get:
      operationId: listMunicipalSuccesses
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - {
            in: query,
            name: districtId,
            schema: { type: string, format: uuid },
          }
        - {
            in: query,
            name: customerType,
            schema: { $ref: "#/components/schemas/CustomerType" },
          }
        - {
            in: query,
            name: attitude,
            schema: { $ref: "#/components/schemas/CustomerAttitude" },
          }
        - {
            in: query,
            name: successOpportunityName,
            schema: { type: string, minLength: 2, maxLength: 100 },
          }
        - {
            in: query,
            name: customerContact,
            schema: { type: string, maxLength: 200 },
          }
        - {
            in: query,
            name: closedFrom,
            schema: { type: string, format: date-time },
          }
        - {
            in: query,
            name: closedTo,
            schema: { type: string, format: date-time },
          }
      responses:
        {
          "200": { description: successful opportunities },
          default: { $ref: "#/components/responses/Problem" },
        }
  /municipal/successes/{opportunityId}:
    get:
      operationId: getMunicipalSuccess
      parameters: [{ $ref: "#/components/parameters/OpportunityId" }]
      responses:
        {
          "200": { description: read-only detail and timeline },
          default: { $ref: "#/components/responses/Problem" },
        }
  /municipal/exports/successes:
    post:
      operationId: exportSuccesses
      parameters: [{ $ref: "#/components/parameters/ReauthToken" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/SuccessExportFilter" },
                },
            },
        }
      responses:
        {
          "200": { $ref: "#/components/responses/Xlsx" },
          default: { $ref: "#/components/responses/Problem" },
        }
  /municipal/exports/workflow:
    post:
      operationId: exportWorkflowNodes
      parameters: [{ $ref: "#/components/parameters/ReauthToken" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/WorkflowExportFilter" },
                },
            },
        }
      responses:
        {
          "200": { $ref: "#/components/responses/Xlsx" },
          default: { $ref: "#/components/responses/Problem" },
        }
  /admin/users:
    get:
      operationId: listUsers
      responses:
        {
          "200": { description: accounts and grants },
          default: { $ref: "#/components/responses/Problem" },
        }
    post:
      operationId: createUser
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/CreateUserRequest" } },
            },
        }
      responses:
        {
          "201": { description: created },
          default: { $ref: "#/components/responses/Problem" },
        }
  /admin/users/{userId}/grants:
    post:
      operationId: createRoleGrant
      parameters:
        [
          {
            in: path,
            name: userId,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: "#/components/schemas/CreateRoleGrantRequest" },
                },
            },
        }
      responses:
        {
          "201": { description: created },
          default: { $ref: "#/components/responses/Problem" },
        }
  /admin/districts/{districtId}/routing:
    post:
      operationId: upsertDistrictRouting
      parameters:
        [
          {
            in: path,
            name: districtId,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/RoutingRequest" } },
            },
        }
      responses:
        {
          "201": { description: saved },
          default: { $ref: "#/components/responses/Problem" },
        }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: oms_session }
    roleGrant: { type: apiKey, in: header, name: X-Role-Grant-Id }
  parameters:
    OpportunityId:
      {
        in: path,
        name: opportunityId,
        required: true,
        schema: { type: string, format: uuid },
      }
...
    ReauthToken:
      {
        in: header,
        name: X-Reauth-Token,
        required: true,
        schema: { type: string },
      }
    Page:
      {
        in: query,
        name: page,
        schema: { type: integer, minimum: 1, default: 1 },
      }
    PageSize:
      {
        in: query,
        name: pageSize,
        schema: { type: integer, minimum: 1, maximum: 100, default: 20 },
      }
  responses:
    Problem:
      description: problem details
      content:
        {
          application/problem+json:
            { schema: { $ref: "#/components/schemas/ProblemDetails" } },
        }
...
    Xlsx:
      description: streamed XLSX without server-side retained copy
      content:
        {
          application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
            { schema: { type: string, format: binary } },
        }
  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required: [phone, password]
      properties:
        {
          phone: { type: string, pattern: '^1[3-9]\d{9}$' },
          password: { type: string, minLength: 8, maxLength: 128 },
        }
    ChangePasswordRequest:
      type: object
      additionalProperties: false
      required: [currentPassword, newPassword]
      properties:
        {
          currentPassword: { type: string, minLength: 8 },
          newPassword: { type: string, minLength: 12, maxLength: 128 },
        }
    ReauthRequest:
      type: object
      additionalProperties: false
      required: [password]
      properties: { password: { type: string, minLength: 8 } }
    RoleCode:
      {
        type: string,
        enum:
          [
            FIELD_REPORTER,
            DISTRICT_MANAGER,
            PERSONAL_HANDLER,
            ORGANIZATION_HANDLER,
            MUNICIPAL,
            SYSTEM_ADMIN,
          ],
      }
    CustomerType: { type: string, enum: [PERSONAL, ORGANIZATION] }
    CustomerAttitude:
      { type: string, enum: [URGENT, IMPORTANT, GENERAL, POTENTIAL] }
    OpportunityState:
      {
        type: string,
        enum:
          [
            PENDING_FIRST_REVIEW,
            RETURNED_TO_REPORTER,
            HANDLING,
            PAUSED,
            PENDING_FINAL_REVIEW,
            RETURNED_TO_HANDLER,
            CLOSED_SUCCESS,
            CLOSED_FAILURE,
          ],
      }
...
    RoleGrant:
      type: object
      additionalProperties: false
      required: [id, role, districtId, districtName]
      properties:
        {
          id: { type: string, format: uuid },
          role: { $ref: "#/components/schemas/RoleCode" },
          districtId: { type: [string, "null"], format: uuid },
          districtName: { type: [string, "null"] },
        }
    CurrentUser:
      type: object
      additionalProperties: false
      required: [id, phone, displayName, mustChangePassword, grants]
      properties:
        {
          id: { type: string, format: uuid },
          phone: { type: string },
          displayName: { type: string },
          mustChangePassword: { type: boolean },
          grants:
            { type: array, items: { $ref: "#/components/schemas/RoleGrant" } },
        }
...
    SuccessExportFilter:
      type: object
      additionalProperties: false
      properties:
        districtId: { type: string, format: uuid }
        customerType: { $ref: "#/components/schemas/CustomerType" }
        attitude: { $ref: "#/components/schemas/CustomerAttitude" }
        successOpportunityName: { type: string, minLength: 2, maxLength: 100 }
        customerContact: { type: string, maxLength: 200 }
        closedFrom: { type: string, format: date-time }
        closedTo: { type: string, format: date-time }
    WorkflowExportFilter:
      type: object
      additionalProperties: false
      properties:
        {
          districtId: { type: string, format: uuid },
          state: { $ref: "#/components/schemas/OpportunityState" },
          submittedFrom: { type: string, format: date-time },
          submittedTo: { type: string, format: date-time },
        }
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [phone, displayName, initialPassword]
      properties:
        {
          phone: { type: string, pattern: '^1[3-9]\d{9}$' },
          displayName: { type: string, minLength: 1, maxLength: 80 },
          initialPassword: { type: string, minLength: 12, maxLength: 128 },
        }
    CreateRoleGrantRequest:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        {
          role: { $ref: "#/components/schemas/RoleCode" },
          districtId: { type: string, format: uuid },
        }
    RoutingRequest:
      type: object
      additionalProperties: false
      required:
        [managerGrantId, personalHandlerGrantId, organizationHandlerGrantId]
      properties:
        {
          managerGrantId: { type: string, format: uuid },
          personalHandlerGrantId: { type: string, format: uuid },
          organizationHandlerGrantId: { type: string, format: uuid },
        }
...
    ProblemDetails:
      type: object
      required: [type, title, status, detail, errorCode]
      properties:
        {
          type: { type: string, format: uri },
          title: { type: string },
          status: { type: integer },
          detail: { type: string },
          instance: { type: string },
          errorCode: { type: string },
          traceId: { type: [string, "null"] },
        }